pyproject.toml
[project]
name = "ankii"
version = "0.1.0"
description = "Anki card reviewer CLI with LLM-powered analysis"
readme = "README.md"
authors = [
    { name = "Mackson Mizumachi", email = "mack.heyha@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "hdbscan>=0.8.41",
    "httpx>=0.28.1",
    "numpy>=2.4.0",
    "plotly>=6.5.0",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "scikit-learn>=1.8.0",
    "streamlit>=1.52.2",
    "typer>=0.21.0",
]

[project.scripts]
ankii = "ankii.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
